Don't allow local override if the sandbox already has one
